# services/polity/openapi/polity.yaml
openapi: 3.0.3
info:
  title: Polity API
  version: 0.1.0
  description: Constitutional government for friend groups.
servers:
  - url: /api/v1
...
paths:
...
  /polities/{polityId}/invitations:
    parameters: [{ $ref: "#/components/parameters/PolityId" }]
    get:
      tags: [Invitations]
      operationId: listPolityInvitations
      parameters: [{ $ref: "#/components/parameters/Page" }, { $ref: "#/components/parameters/PageSize" }]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Membership invitations, content: { application/json: { schema: { $ref: "#/components/schemas/MemberInvitationPageResponse" } } } }
    post:
      tags: [Invitations]
      operationId: createPolityInvitation
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CreateMemberInvitationRequest" } } } }
      responses:
        "201": { description: Membership invitation created, content: { application/json: { schema: { $ref: "#/components/schemas/MemberInvitationResponse" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
  /invitations:
    get:
      tags: [Invitations]
      operationId: listCurrentUserInvitations
      parameters: [{ $ref: "#/components/parameters/Page" }, { $ref: "#/components/parameters/PageSize" }]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Pending invitations for the current user, content: { application/json: { schema: { $ref: "#/components/schemas/MemberInvitationPageResponse" } } } }
  /invitations/{invitationId}/accept:
    parameters: [{ $ref: "#/components/parameters/InvitationId" }]
    post:
      tags: [Invitations]
      operationId: acceptInvitation
      security: [{ bearerAuth: [] }]
      responses:
        "201": { description: Invitation accepted and membership created, content: { application/json: { schema: { $ref: "#/components/schemas/MemberResponse" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /polities/{polityId}/motions:
    parameters: [{ $ref: "#/components/parameters/PolityId" }]
    get:
      tags: [Motions]
      operationId: listPolityMotions
      parameters: [{ $ref: "#/components/parameters/Page" }, { $ref: "#/components/parameters/PageSize" }]
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Motions, content: { application/json: { schema: { $ref: "#/components/schemas/MotionPageResponse" } } } }
    post:
      tags: [Motions]
      operationId: createPolityMotion
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CreateMotionRequest" } } } }
      responses:
        "201": { description: Motion introduced and opened for voting, content: { application/json: { schema: { $ref: "#/components/schemas/MotionResponse" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
  /polities/{polityId}/motions/office-elections:
    parameters: [{ $ref: "#/components/parameters/PolityId" }]
    post:
      tags: [Motions]
      operationId: createPolityOfficeElectionMotion
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CreateOfficeElectionMotionRequest" } } } }
      responses:
        "201": { description: Office election motion introduced, content: { application/json: { schema: { $ref: "#/components/schemas/MotionResponse" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /polities/{polityId}/motions/sanctions:
    parameters: [{ $ref: "#/components/parameters/PolityId" }]
    post:
      tags: [Motions]
      operationId: createPolitySanctionMotion
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CreateSanctionMotionRequest" } } } }
      responses:
        "201": { description: Sanction motion introduced, content: { application/json: { schema: { $ref: "#/components/schemas/MotionResponse" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /polities/{polityId}/motions/appeals:
    parameters: [{ $ref: "#/components/parameters/PolityId" }]
    post:
      tags: [Motions]
      operationId: createPolityAppealMotion
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CreateAppealMotionRequest" } } } }
      responses:
        "201": { description: Appeal motion introduced, content: { application/json: { schema: { $ref: "#/components/schemas/MotionResponse" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /polities/{polityId}/motions/office-term-reviews:
    parameters: [{ $ref: "#/components/parameters/PolityId" }]
    post:
      tags: [Motions]
      operationId: createPolityOfficeTermReviewMotion
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CreateOfficeTermReviewMotionRequest" } } } }
      responses:
        "201": { description: Office term review motion introduced, content: { application/json: { schema: { $ref: "#/components/schemas/MotionResponse" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /polities/{polityId}/motions/constitutional-reviews:
    parameters: [{ $ref: "#/components/parameters/PolityId" }]
    post:
      tags: [Motions]
      operationId: createPolityConstitutionalReviewMotion
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CreateConstitutionalReviewMotionRequest" } } } }
      responses:
        "201": { description: Constitutional review motion introduced, content: { application/json: { schema: { $ref: "#/components/schemas/MotionResponse" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /polities/{polityId}/motions/constitution-amendments:
    parameters: [{ $ref: "#/components/parameters/PolityId" }]
    post:
      tags: [Motions]
      operationId: createPolityConstitutionAmendmentMotion
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CreateConstitutionAmendmentMotionRequest" } } } }
      responses:
        "201": { description: Constitution amendment motion introduced, content: { application/json: { schema: { $ref: "#/components/schemas/MotionResponse" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
  /polities/{polityId}/motions/disbandment:
    parameters: [{ $ref: "#/components/parameters/PolityId" }]
    post:
      tags: [Motions]
      operationId: createPolityDisbandmentMotion
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CreateDisbandmentMotionRequest" } } } }
      responses:
        "201": { description: Disbandment motion introduced, content: { application/json: { schema: { $ref: "#/components/schemas/MotionResponse" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
  /polities/{polityId}/motions/{motionId}:
    parameters: [{ $ref: "#/components/parameters/PolityId" }, { $ref: "#/components/parameters/MotionId" }]
    get:
      tags: [Motions]
      operationId: getPolityMotion
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Motion, content: { application/json: { schema: { $ref: "#/components/schemas/MotionResponse" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
  /polities/{polityId}/motions/{motionId}/votes:
    parameters: [{ $ref: "#/components/parameters/PolityId" }, { $ref: "#/components/parameters/MotionId" }]
    put:
      tags: [Motions]
      operationId: castPolityMotionVote
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CastVoteRequest" } } } }
      responses:
        "200": { description: Vote recorded, content: { application/json: { schema: { $ref: "#/components/schemas/MotionResponse" } } } }
        "409": { $ref: "#/components/responses/Conflict" }
  /polities/{polityId}/motions/{motionId}/election-ballot:
    parameters: [{ $ref: "#/components/parameters/PolityId" }, { $ref: "#/components/parameters/MotionId" }]
    put:
      tags: [Motions]
      operationId: castPolityOfficeElectionBallot
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CastOfficeElectionBallotRequest" } } } }
      responses:
        "200": { description: Election ballot recorded, content: { application/json: { schema: { $ref: "#/components/schemas/MotionResponse" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /polities/{polityId}/motions/{motionId}/candidacy:
    parameters: [{ $ref: "#/components/parameters/PolityId" }, { $ref: "#/components/parameters/MotionId" }]
    put:
      tags: [Motions]
      operationId: respondPolityOfficeElectionCandidacy
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/RespondOfficeElectionCandidacyRequest" } } } }
      responses:
        "200": { description: Candidacy response recorded, content: { application/json: { schema: { $ref: "#/components/schemas/MotionResponse" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /polities/{polityId}/motions/{motionId}/certification:
    parameters: [{ $ref: "#/components/parameters/PolityId" }, { $ref: "#/components/parameters/MotionId" }]
    post:
      tags: [Motions]
      operationId: certifyPolityMotion
      security: [{ bearerAuth: [] }]
      responses:
        "201": { description: Result certified and effect applied when passed, content: { application/json: { schema: { $ref: "#/components/schemas/MotionResponse" } } } }
        "409": { $ref: "#/components/responses/Conflict" }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    PolityId: { name: polityId, in: path, required: true, schema: { type: string, format: uuid } }
    MotionId: { name: motionId, in: path, required: true, schema: { type: string, format: uuid } }
    InvitationId: { name: invitationId, in: path, required: true, schema: { type: string, format: uuid } }
...
    Page:
      name: page
      in: query
      required: false
      description: Zero-based page index.
      schema: { type: integer, minimum: 0, default: 0 }
    PageSize:
      name: size
      in: query
      required: false
      description: Number of items per page.
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
  responses:
    BadRequest:
      description: Invalid request.
      content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
    Forbidden:
      description: The caller is not authorized for this action.
      content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
    NotFound:
      description: Resource not found.
      content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
    Conflict:
      description: Resource conflicts with current state.
      content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
  schemas:
    ApiError:
      type: object
      required: [error]
      properties:
        error: { $ref: "#/components/schemas/ApiErrorBody" }
    ApiErrorBody:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
    PageMetadata:
      type: object
      required: [size, number, totalElements, totalPages]
      properties:
        size: { type: integer, format: int64, minimum: 1 }
        number: { type: integer, format: int64, minimum: 0 }
        totalElements: { type: integer, format: int64, minimum: 0 }
        totalPages: { type: integer, format: int64, minimum: 0 }
...
    MemberInvitationPageResponse:
      type: object
      required: [content, page]
      properties:
        content: { type: array, items: { $ref: "#/components/schemas/MemberInvitationResponse" } }
        page: { $ref: "#/components/schemas/PageMetadata" }
    MotionPageResponse:
      type: object
      required: [content, page]
      properties:
        content: { type: array, items: { $ref: "#/components/schemas/MotionResponse" } }
        page: { $ref: "#/components/schemas/PageMetadata" }
...
    CreateMemberInvitationRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email, x-field-extra-annotation: "@jakarta.validation.constraints.NotBlank" }
    CreateMotionRequest:
      type: object
      required: [title, body]
      properties:
        title: { type: string, minLength: 1, maxLength: 200, x-field-extra-annotation: "@jakarta.validation.constraints.NotBlank" }
        body: { type: string, minLength: 1, maxLength: 5000, x-field-extra-annotation: "@jakarta.validation.constraints.NotBlank" }
    CreateDisbandmentMotionRequest:
      type: object
      required: [title, body]
      properties:
        title: { type: string, minLength: 1, maxLength: 200, x-field-extra-annotation: "@jakarta.validation.constraints.NotBlank" }
        body: { type: string, minLength: 1, maxLength: 5000, x-field-extra-annotation: "@jakarta.validation.constraints.NotBlank" }
    CreateOfficeElectionMotionRequest:
      type: object
      required: [officeId, candidateMembershipIds]
      properties:
        officeId: { type: string, format: uuid }
        candidateMembershipIds:
          type: array
          minItems: 1
          maxItems: 12
          uniqueItems: true
          items: { type: string, format: uuid }
    CreateSanctionMotionRequest:
      type: object
      required: [targetMembershipId, type, reason, durationDays]
      properties:
        targetMembershipId: { type: string, format: uuid }
        type: { $ref: "#/components/schemas/SanctionType" }
        reason: { type: string, minLength: 1, maxLength: 5000, x-field-extra-annotation: "@jakarta.validation.constraints.NotBlank" }
        durationDays:
          type: integer
          minimum: 1
          description: Must be long enough for the current appeal procedure to complete.
    CreateAppealMotionRequest:
      type: object
      required: [sanctionId, reason]
      properties:
        sanctionId: { type: string, format: uuid }
        reason: { type: string, minLength: 1, maxLength: 5000, x-field-extra-annotation: "@jakarta.validation.constraints.NotBlank" }
    CreateOfficeTermReviewMotionRequest:
      type: object
      required: [officeTermId, reason]
      properties:
        officeTermId: { type: string, format: uuid }
        reason: { type: string, minLength: 1, maxLength: 5000, x-field-extra-annotation: "@jakarta.validation.constraints.NotBlank" }
    CreateConstitutionalReviewMotionRequest:
      type: object
      description: Opens a constitutional-review motion for an official record with an active void remedy.
      required: [targetRecordId, reason]
      properties:
        targetRecordId:
          type: string
          format: uuid
          description: Official record entry to void. The entry must be a resolution adoption, sanction application, or office election with an active remedy.
        reason: { type: string, minLength: 1, maxLength: 5000, x-field-extra-annotation: "@jakarta.validation.constraints.NotBlank" }
    CreateConstitutionAmendmentMotionRequest:
      type: object
      required: [title, body]
      properties:
        title: { type: string, minLength: 1, maxLength: 200, x-field-extra-annotation: "@jakarta.validation.constraints.NotBlank" }
        body: { type: string, minLength: 1, maxLength: 10000, x-field-extra-annotation: "@jakarta.validation.constraints.NotBlank" }
        institutionChanges:
          type: array
          maxItems: 10
          items: { $ref: "#/components/schemas/CreateInstitutionChangeRequest" }
        procedureChanges:
          type: array
          maxItems: 10
          items: { $ref: "#/components/schemas/CreateProcedureChangeRequest" }
        officeChanges:
          type: array
          maxItems: 10
          items: { $ref: "#/components/schemas/CreateOfficeChangeRequest" }
        powerChanges:
          type: array
          maxItems: 10
          items: { $ref: "#/components/schemas/CreatePowerChangeRequest" }
    CreateInstitutionChangeRequest:
      type: object
      required: [action]
      properties:
        action: { $ref: "#/components/schemas/ConstitutionInstitutionChangeAction" }
        institutionId: { type: string, format: uuid }
        jurisdictionId: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 120 }
        kind: { $ref: "#/components/schemas/InstitutionKind" }
    CreateProcedureChangeRequest:
      type: object
      required: [procedureCode]
      properties:
        procedureCode: { type: string, minLength: 1, x-field-extra-annotation: "@jakarta.validation.constraints.NotBlank" }
        institutionId: { type: string, format: uuid }
        quorumNumerator: { type: integer, minimum: 1 }
        quorumDenominator: { type: integer, minimum: 1 }
        threshold: { $ref: "#/components/schemas/VotingThreshold" }
        officeElectionMethod: { $ref: "#/components/schemas/OfficeElectionMethod" }
        electorate: { $ref: "#/components/schemas/ProcedureElectorate" }
        electorateOfficeCode: { type: string, minLength: 1, maxLength: 64, pattern: "^[a-z][a-z0-9-]*$" }
        minimumElectorCount: { type: integer, minimum: 1 }
        minimumNoticeHours: { type: integer, minimum: 0 }
        votingPeriodHours: { type: integer, minimum: 1 }
...
    CastVoteRequest:
      type: object
      required: [choice]
      properties:
        choice: { $ref: "#/components/schemas/VoteChoice" }
    CastOfficeElectionBallotRequest:
      type: object
      required: [candidateMembershipIds]
      properties:
        candidateMembershipIds:
          type: array
          minItems: 1
          maxItems: 12
          uniqueItems: true
          items: { type: string, format: uuid }
    RespondOfficeElectionCandidacyRequest:
      type: object
      required: [accepted]
      properties:
        accepted: { type: boolean }
    VoteChoice: { type: string, enum: ["yes", "no", abstain] }
...
    InstitutionKind: { type: string, enum: [assembly, council, judiciary] }
...
    InvitationStatus: { type: string, enum: [pending, accepted] }
...
    SanctionType: { type: string, enum: [warning, suspension] }
...
    EffectType: { type: string, enum: [adopt_resolution, elect_office, apply_sanction, grant_appeal, vacate_office_term, void_official_act, amend_constitution, disband_polity] }
...
    MemberResponse:
      type: object
      required: [id, userId, name, email, status, admittedAt]
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email }
        status: { type: string, enum: [active, resigned] }
        admittedAt: { type: string, format: date-time }
    MemberInvitationResponse:
      type: object
      required: [id, polityId, polityName, email, invitedByName, status, invitedAt]
      properties:
        id: { type: string, format: uuid }
        polityId: { type: string, format: uuid }
        polityName: { type: string }
        email: { type: string, format: email }
        invitedByName: { type: string }
        status: { $ref: "#/components/schemas/InvitationStatus" }
        invitedAt: { type: string, format: date-time }
        respondedAt: { type: string, format: date-time }
    VoteTallyResponse:
      type: object
      required: ["yes", "no", abstain, eligible, quorumRequired, quorumMet, outcomeReason]
      properties:
        "yes": { type: integer }
        "no": { type: integer }
        abstain: { type: integer }
        eligible: { type: integer }
        quorumRequired: { type: integer }
        quorumMet: { type: boolean }
        outcomeReason: { $ref: "#/components/schemas/VotingOutcomeReason" }
    CertificationResponse:
      type: object
      description: Durable evidence captured when the motion was certified. Modality-specific counts are present only for their corresponding modality.
      required: [modality, eligibleCount, quorumRequired, quorumMet, thresholdMet, passed, outcomeReason, certifiedAt]
      properties:
        modality: { $ref: "#/components/schemas/CertificationModality" }
        eligibleCount: { type: integer }
        yesCount: { type: integer }
        noCount: { type: integer }
        abstainCount: { type: integer }
        electionParticipationCount: { type: integer }
        electionDecisive: { type: boolean }
        electionWinnerCount: { type: integer }
        quorumRequired: { type: integer }
        quorumMet: { type: boolean }
        thresholdMet: { type: boolean }
        passed: { type: boolean }
        outcomeReason: { $ref: "#/components/schemas/CertificationOutcomeReason" }
        certifiedAt: { type: string, format: date-time }
...
    MotionResponse:
      type: object
      required: [id, title, body, status, effectType, constitutionVersion, procedureName, introducedByName, openedAt, votingOpensAt, votingClosesAt, certificationOpensAt, actions]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        body: { type: string }
        titleKey: { type: string }
        bodyKey: { type: string }
        templateParams:
          type: object
          additionalProperties: true
        status: { type: string, enum: [voting, enacted, rejected] }
        effectType: { $ref: "#/components/schemas/EffectType" }
        constitutionVersion: { type: integer }
        procedureName: { type: string }
        procedureNameKey: { type: string }
        introducedByName: { type: string }
        openedAt: { type: string, format: date-time }
        votingOpensAt: { type: string, format: date-time }
        votingClosesAt: { type: string, format: date-time }
        certificationOpensAt: { type: string, format: date-time }
        tally:
          allOf: [{ $ref: "#/components/schemas/VoteTallyResponse" }]
          description: Live tally before certification; the durable certified tally after certification.
        officeElection: { $ref: "#/components/schemas/OfficeElectionResponse" }
        electionTally:
          allOf: [{ $ref: "#/components/schemas/OfficeElectionTallyResponse" }]
          description: Live election tally before certification; the durable certified snapshot after certification.
        certification: { $ref: "#/components/schemas/CertificationResponse" }
        currentVote: { $ref: "#/components/schemas/VoteChoice" }
        actions: { $ref: "#/components/schemas/MotionActionAvailabilityResponse" }
        amendmentProposal: { $ref: "#/components/schemas/ConstitutionAmendmentProposalResponse" }
    MotionActionAvailabilityResponse:
      type: object
      required: [castVote, castElectionBallot, respondCandidacy, requestCertification]
      properties:
        castVote: { $ref: "#/components/schemas/ActionAvailabilityResponse" }
        castElectionBallot: { $ref: "#/components/schemas/ActionAvailabilityResponse" }
        respondCandidacy: { $ref: "#/components/schemas/ActionAvailabilityResponse" }
        requestCertification: { $ref: "#/components/schemas/ActionAvailabilityResponse" }
    ConstitutionAmendmentProposalResponse:
      type: object
      required: [title, body, institutionChanges, procedureChanges, officeChanges, powerChanges]
      properties:
        title: { type: string }
        body: { type: string }
        institutionChanges:
          type: array
          items: { $ref: "#/components/schemas/ConstitutionInstitutionChangeResponse" }
        procedureChanges:
          type: array
          items: { $ref: "#/components/schemas/ConstitutionProcedureChangeResponse" }
        officeChanges:
          type: array
          items: { $ref: "#/components/schemas/ConstitutionOfficeChangeResponse" }
        powerChanges:
          type: array
          items: { $ref: "#/components/schemas/ConstitutionPowerChangeResponse" }
...
    OfficeElectionResponse:
      type: object
      required: [officeId, officeCode, officeName, seatsAvailable, method, candidates]
      properties:
        officeId: { type: string, format: uuid }
        officeCode: { type: string }
        officeName: { type: string }
        officeNameKey: { type: string }
        seatsAvailable: { type: integer }
        method: { $ref: "#/components/schemas/OfficeElectionMethod" }
        currentBallot: { $ref: "#/components/schemas/OfficeElectionBallotResponse" }
        candidates:
          type: array
          items: { $ref: "#/components/schemas/OfficeElectionCandidateResponse" }
...
    OfficeElectionTallyResponse:
      type: object
      required: [eligible, participation, quorumRequired, quorumMet, seatsAvailable, seatsFilled, method, decisive, passed, outcomeReason, winners, candidates, rounds]
      properties:
        eligible: { type: integer }
        participation: { type: integer }
        quorumRequired: { type: integer }
        quorumMet: { type: boolean }
        seatsAvailable: { type: integer }
        seatsFilled: { type: integer }
        method: { $ref: "#/components/schemas/OfficeElectionMethod" }
        quota: { type: number }
        decisive: { type: boolean }
        passed: { type: boolean }
        outcomeReason: { $ref: "#/components/schemas/OfficeElectionOutcomeReason" }
        winners:
          type: array
          items: { $ref: "#/components/schemas/OfficeElectionCandidateTallyResponse" }
        candidates:
          type: array
          items: { $ref: "#/components/schemas/OfficeElectionCandidateTallyResponse" }
        rounds:
          type: array
          items: { $ref: "#/components/schemas/OfficeElectionRoundResponse" }
...